+++ src/AnxPro.Api/AnxProClient.Registration.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using AnxPro.Api.Models;
 
@@ -12,15 +13,25 @@ namespace AnxPro.Api
         /// <param name="username">User account name.  E.g., user's email.</param>
         /// <param name="userTimezone">Timezone of user's registration country.  E.g., Hong Kong.</param>
         /// <param name="country">Three letter country code.</param>
-        /// <param name="accepttc">True to accept t&c of the platform. Must set to true to register successfully.</param>
+        /// <param name="acceptTc">True to accept t&c of the platform. Must set to true to register successfully.</param>
         /// <param name="language">User preferred language. E.g., en</param>
         /// <remarks>
         /// The platform will send email with verification code if user registered correctly.
         /// </remarks>
         /// <returns></returns>
-        public async Task<AnxProResponse> RegisterAsync(string username, string userTimezone, string country, bool accepttc, string language)
+        public async Task<AnxProResponse> RegisterAsync(string username, string userTimezone, string country, bool acceptTc, string language)
         {
-            throw new NotImplementedException();
+            return await QueryPrivateAsync<AnxProResponse>(
+                "register/register",
+                new Dictionary<string, string>(5)
+                {
+                    ["username"] = username,
+                    ["userTimezone"] = userTimezone,
+                    ["country"] = country,
+                    ["accepttc"] = acceptTc ? "1" : "0",
+                    ["language"] = language,
+                }
+            );
         }
 
         /// <summary>
@@ -29,7 +40,7 @@ namespace AnxPro.Api
         /// <param name="verificationCode">Verification code.</param>
         /// <param name="password">Password.</param>
         /// <param name="password2"><c>Password2</c> needs to match <c>Password</c>.</param>
-        /// <param name="accepttc">True to accept t&c of the platform. Must set to true to register successfully.</param>
+        /// <param name="acceptTc">True to accept t&c of the platform. Must set to true to register successfully.</param>
         /// <param name="country">Three letter country code.</param>
         /// <remarks>
         /// No Rest-Key and Rest-Sign in HTTP header.
@@ -38,9 +49,19 @@ namespace AnxPro.Api
         /// <returns>
         /// <c>ResultCode</c> <c>OK</c> if verification succeeded. <c>SYSTEM_ERROR</c> if verification failed.
         /// </returns>
-        public async Task<AnxProResponse> VerifyRegistrationAsync(string verificationCode, string password, string password2, bool accepttc, string country)
+        public async Task<AnxProResponse> VerifyRegistrationAsync(string verificationCode, string password, string password2, bool acceptTc, string country)
         {
-            throw new NotImplementedException();
+            return await QueryPrivateAsync<AnxProResponse>(
+                "register/verifyRegistration",
+                new Dictionary<string, string>(5)
+                {
+                    ["verificationCode"] = verificationCode,
+                    ["password"] = password,
+                    ["password2"] = password2,
+                    ["accepttc"] = acceptTc ? "1" : "0",
+                    ["country"] = country,
+                }
+            );
         }
 
         /// <summary>
@@ -56,7 +77,13 @@ namespace AnxPro.Api
         /// </returns>
         public async Task<AnxProResponse> ResendVerificationAsync(string username)
         {
-            throw new NotImplementedException();
+            return await QueryPrivateAsync<AnxProResponse>(
+                "register/resendVerification",
+                new Dictionary<string, string>(1)
+                {
+                    ["username"] = username,
+                }
+            );
         }
 
         /// <summary>
@@ -64,7 +91,7 @@ namespace AnxPro.Api
         /// </summary>
         /// <param name="userTimezone">Timezone of user's registration country.</param>
         /// <param name="country">Three letter country code.</param>
-        /// <param name="accepttc">True to accept t&c of the platform. Must set to true to register successfully.</param>
+        /// <param name="acceptTc">True to accept t&c of the platform. Must set to true to register successfully.</param>
         /// <param name="language">User preferred language.</param>
         /// <param name="usernameType">Type of user account name. E.g., PHONENUMBER</param>
         /// <param name="username">User account name. E.g., +85295271816</param>
@@ -72,9 +99,20 @@ namespace AnxPro.Api
         /// This account cannot be login from website/mobile app. The account is only accessible by the master key api.
         /// </remarks>
         /// <returns></returns>
-        public async Task<AutoRegisterResponse> AutoRegisterAsync(string userTimezone, string country, bool accepttc, Language language, UserNameType usernameType, string username)
+        public async Task<AutoRegisterResponse> AutoRegisterAsync(string userTimezone, string country, bool acceptTc, Language language, UserNameType usernameType, string username)
         {
-            throw new NotImplementedException();
+            return await QueryPrivateAsync<AutoRegisterResponse>(
+                "register/autoRegister",
+                new Dictionary<string, string>(6)
+                {
+                    ["userTimezone"] = userTimezone,
+                    ["country"] = country,
+                    ["accepttc"] = acceptTc ? "1" : "0",
+                    ["language"] = language.ToString(),
+                    ["usernameType"] = usernameType.ToString(),
+                    ["username"] = username,
+                }
+            );
         }
     }
 }